Divided management commands on one file
